Split attestation types into sgx and sev_snp specific headers
